Fix typo
Fixed the opening sentence. As is, I don't believe it's grammatical. 

You could also do something like this: "This a guide (something between a tiny book and a course) that shows new users how to get started with Clojure."

Or even this: "This a guide that shows new users how to get started with Clojure. It is brief and focussed: something between a tiny book and a course."